coral/object_detection_app/requirements.txt
numpy
pynetworktables
python3-opencv
tflite-runtime==2.11.0